Bugfix: HSV recolor rounding Value/Saturation unintentionally
Fixed an issue where the HSV recolor method would round value and saturation. This means that Functions and Easings now return floats instead of ints, and rounding will need to be done on the user code side if they make use of the easings directly.

+++ AnimateForms/Animate/Animate.cs
@@ -10,7 +10,7 @@ namespace AnimateForms.Animate
 {
     public class Animate
     {
-        public delegate int Function(float t, float b, float c, float d);
+        public delegate float Function(float t, float b, float c, float d);
         private readonly List<(string, string)> _animating = new List<(string, string)>();
 
         public async Task<bool> Move(Control control, Function easing, int duration, Point moveTo)
@@ -37,8 +37,8 @@ namespace AnimateForms.Animate
                 int time = (int)stopwatch.ElapsedMilliseconds;
                 if (time > duration) time = duration;
 
-                control.Location = new Point(easing(time, location.X, xDif, duration),
-                                             easing(time, location.Y, yDif, duration));
+                control.Location = new Point((int)easing(time, location.X, xDif, duration),
+                                             (int)easing(time, location.Y, yDif, duration));
             }
 
             _animating.Remove((control.Name, "move"));
@@ -97,8 +97,8 @@ namespace AnimateForms.Animate
                 int time = (int)stopwatch.ElapsedMilliseconds;
                 if (time > duration) time = duration;
 
-                Point newPoint = new Point(easing(time, 0, offset.X, duration),
-                                           easing(time, 0, offset.Y, duration));
+                Point newPoint = new Point((int)easing(time, 0, offset.X, duration),
+                                           (int)easing(time, 0, offset.Y, duration));
 
                 control.Location = new Point(control.Location.X + (newPoint.X - prevPoint.X),
                                              control.Location.Y + (newPoint.Y - prevPoint.Y));
@@ -156,10 +156,10 @@ namespace AnimateForms.Animate
                 if (time > duration) time = duration;
 
                 Color newColor;
-                newColor = Color.FromArgb(easing(time, color.A, aDif, duration),
-                                          easing(time, color.R, rDif, duration),
-                                          easing(time, color.G, gDif, duration),
-                                          easing(time, color.B, bDif, duration));
+                newColor = Color.FromArgb((int)easing(time, color.A, aDif, duration),
+                                          (int)easing(time, color.R, rDif, duration),
+                                          (int)easing(time, color.G, gDif, duration),
+                                          (int)easing(time, color.B, bDif, duration));
                 if (backColor)
                     control.BackColor = newColor;
                 else
@@ -189,8 +189,8 @@ namespace AnimateForms.Animate
             }
 
             int hDif = (int)(colorTo.Hue - color.Hue);
-            int sDif = (int)(colorTo.Saturation - color.Saturation);
-            int vDif = (int)(colorTo.Value - color.Value);
+            float sDif = colorTo.Saturation - color.Saturation;
+            float vDif = colorTo.Value - color.Value;
 
             Stopwatch stopwatch = new Stopwatch();
             stopwatch.Start();
@@ -314,8 +314,8 @@ namespace AnimateForms.Animate
                 int time = (int)stopwatch.ElapsedMilliseconds;
                 if (time > duration) time = duration;
 
-                control.Size = new Size(easing(time, size.Width, widthDif, duration),
-                                        easing(time, size.Height, heightDif, duration));
+                control.Size = new Size((int)easing(time, size.Width, widthDif, duration),
+                                        (int)easing(time, size.Height, heightDif, duration));
             }
 
             _animating.Remove((control.Name, "resize"));
